requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Cython==0.29.21
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-js-asset==1.2.2
django-mptt==0.11.0
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.1
numpy==1.19.4
Pillow==7.2.0
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.11
whitenoise==5.2.0
